"Three Good Things" is an android application for doing a positive psychology exercise.
The app allows the user to keep a diary of three good things that he/she
experienced a particular day.
The date of the day and the three good things of the day are stored
permanently in an SQLite database.

The app allows the user to add an entry to the diary, but also to update
an entry (change good things/date) or even delete an entry.

The app allows only one entry for a particular date.